Create README.md
This project contains a game of Rock-Paper-Scissors with use of simplest Python statements. 
I have used different functions to make it easier. In the first function "get_user_choice" this function takes a user input from a given list of "rock, paper and scissors". User can choose any one of this.
The next function "get_computer_choice" uses random function with a range of "rock, paper and scissors". So now we are ready with both the user input and a computer choice.
The next function "determine_winner" tests if the user defeats the computer in the game or the computer defeats the user.
At last the "play_game" function acts as a main function which calls the above functions to display the final output!